backend/requirements.txt
Django==3.2.8
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-crontab==0.7.1
django-environ==0.9.0